Render a clerk_tbl to its target output format

Dispatches to the correct renderer based on the output slot set at construction time (tbl_descriptive(..., output = "gt"|"html"|"latex")). All render arguments (title, subtitle, footnote) can be supplied here and are forwarded to the underlying renderer.

Usage

clerk_render(x, title = NULL, subtitle = NULL, footnote = NULL, ...)

Arguments

x

A clerk_tbl object.

title

Optional character string for the table title.

subtitle

Optional character string for the table subtitle.

footnote

Optional additional footnote text.

...

Passed to the underlying render_gt(), render_reactable(), or render_latex().

Value

A gt_tbl, reactable, or knit_asis object depending on the output slot of x.

Examples

tbl_descriptive(clerk_example, group = sex, output = "gt") |>
  clerk_render(title = "Table 1")

  

Table 1
    

      	n
      	Overall
      	Female
      	Male
      	Statistic
      	p
    

All variables
    

age
	300
	46.76 ± 14.15
	46.42 ± 13.63
	47.35 ± 15.09
	t = -0.53
	5.992006e-01

hdl
	300
	46.81 ± 10.11
	48.32 ± 9.78
	44.08 ± 10.18
	t = 3.50
	5.599878e-04

glucose
	300
	91.68 ± 24.65
	92.61 ± 24.49
	90.01 ± 24.96
	t = 0.87
	3.863309e-01

bmi
	300
	26.15 ± 4.91
	26.43 ± 4.84
	25.64 ± 5.01
	t = 1.32
	1.895409e-01

waist
	300
	92.07 ± 11.56
	92.14 ± 11.32
	91.93 ± 12.03
	t = 0.15
	8.792922e-01

systolic_bp
	300
	123.25 ± 17.77
	120.36 ± 17.86
	128.48 ± 16.45
	t = -3.97
	9.521407e-05

tmt_time
	300
	129.24 ± 64.98
	132.81 ± 66.77
	122.81 ± 61.40
	t = 1.31
	1.916489e-01

verbal_fluency
	300
	14.98 ± 4.63
	15.08 ± 4.44
	14.79 ± 4.97
	t = 0.52
	6.062013e-01

bdi
	300
	13.64 ± 9.26
	15.07 ± 8.97
	11.05 ± 9.26
	t = 3.65
	3.325514e-04

panas_neg
	300
	20.98 ± 7.48
	22.50 ± 7.27
	18.25 ± 7.09
	t = 4.92
	1.662527e-06

life_satisfaction
	300
	18.02 ± 4.23
	17.40 ± 4.25
	19.14 ± 3.99
	t = -3.54
	4.880922e-04

tbl_descriptive(clerk_example, group = sex, output = "html") |>
  clerk_render()